chore(typing): Clear the mechanical mypy findings, 29 -> 1
Three kinds of noise, none of them a behavior change:

- Empty containers mypy cannot infer from later appends: label_counts,
  diag_warnings, and the np.zeros/np.ones buffers whose dtype argument leaves
  the element type unsolved.

- numpy scalars bound to names the code then treats as Python numbers. Wrapping
  at the assignment (int()/float()) is where the conversion belongs: n_invalid
  and n_negative only ever land in a message, ss_res/ss_tot and neg_sum/pos_sum
  only feed arithmetic, and rank already had an int() bolted on at one of its
  two call sites.

- Return values that had drifted from their annotation: compute_fit_metrics,
  robust_nnls and estimate_R_linear all promise float and were handing back
  floating[Any].

Two are slightly more than cosmetic. zhit.py now passes int(np.argmin(...)) to
a parameter declared int rather than an np.intp. And numpy 2 renamed trapz to
trapezoid, so auto_suggest's 1.x fallback import cannot resolve against the 2.x
stubs - it gets a narrow ignore rather than a restructure, since the fallback
goes away with the numpy floor.

Left standing: model_comparison.py:28, where the ModelScore NamedTuple's
`index` field shadows tuple.index. Renaming it touches every call site, so it
is a change to make deliberately, not in a sweep.

ruff clean, 535 tests pass.

+++ eis_analysis/fitting/diagnostics.py
@@ -92,8 +92,8 @@ def compute_fit_metrics(
     # the residual and the magnitude, so it is not double-counted with the
     # 1/|Z| that already defines a relative error. For modulus weighting
     # (w = 1/|Z|) this reduces to the mean relative error mean(|dZ|/|Z|).
-    fit_error_rel = np.sum(weights * abs_errors) / np.sum(weights * Z_mag_safe) * 100
-    fit_error_abs = np.mean(abs_errors)
+    fit_error_rel = float(np.sum(weights * abs_errors) / np.sum(weights * Z_mag_safe) * 100)
+    fit_error_abs = float(np.mean(abs_errors))
 
     # Log unweighted vs weighted difference if significant
     fit_error_rel_unweighted = np.mean(relative_errors) * 100
@@ -192,7 +192,7 @@ def compute_information_criteria(
     # correct limit - such a model wins every comparison - but -inf poisons
     # the delta arithmetic that follows, so clamp to the smallest positive
     # normal instead and let the ranking stay finite.
-    rss_safe = max(rss, np.finfo(float).tiny)
+    rss_safe = max(rss, float(np.finfo(float).tiny))
 
     log_likelihood_term = n * np.log(rss_safe / n)
     aic = log_likelihood_term + 2.0 * k
